Move the server tests to the tests/server-common module
